requirements.txt
fastapi[all]
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
email-validator
uvicorn[standard]